+++ packages/@overeng/stylex-tokens/tsconfig.json.genie.ts
@@ -9,8 +9,6 @@ export default tsconfigJson({
   compilerOptions: {
     ...baseTsconfigCompilerOptions,
     ...packageTsconfigCompilerOptions,
-    allowJs: true,
-    checkJs: true,
     lib: [...domLib],
   },
   include: ['src/**/*'],